requirements.txt
abstention==0.1.3.1
hydra-core==1.3.2
matplotlib==3.9.2
numpy==2.1.2
omegaconf==2.3.0
opencv-python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
lightning==2.4.0
scikit-learn==1.5.2
scipy==1.14.1
scikit-image==0.0
timm==1.0.11
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
seaborn==0.13.2
